fix: detail statistics accounts data
