+++ packages/next-common/components/statistics/referenda/proposalAddress/accountsVotesContent.jsx
@@ -1,18 +1,21 @@
 import { StatisticsItemDiv } from "./style";
 import AccountsVotesChart from "./accountsVotesChart";
 import { useTheme } from "styled-components";
-import { useRangedVotesData } from "./useRangedVotesData";
-import { useChainSettings } from "next-common/context/chain";
+import { useChain } from "next-common/context/chain";
+import { toRangeLabels, VOTES_RANGE } from "next-common/utils/statistics/range";
+import { useVotesCount } from "next-common/hooks/referenda/useVotesCount";
 
 export default function AccountsVotesContent() {
   const theme = useTheme();
-  const { voteSymbol, symbol } = useChainSettings();
-  const { rangedLabels, ayes, nays, abstains } = useRangedVotesData();
+  const { ayes, nays, abstains } = useVotesCount();
+  const chain = useChain();
+  const range = VOTES_RANGE[chain];
+  const labels = toRangeLabels(range);
 
   const categoryPercentage = 1;
   const barPercentage = 0.1;
   const dataVotes = {
-    labels: rangedLabels,
+    labels,
     datasets: [
       {
         label: "Aye",
@@ -42,7 +45,7 @@ export default function AccountsVotesContent() {
       <div className="flex gap-2 mb-4">
         <span className="text14Bold text-textPrimary">Accounts</span>
       </div>
-      <AccountsVotesChart data={dataVotes} symbol={voteSymbol || symbol} />
+      <AccountsVotesChart data={dataVotes} />
     </StatisticsItemDiv>
   );
 }
